[CANVAS][REFACTO] split code in files
